Reorganize lexer tests: separate proptests from integration tests
This commit reorganizes tests/lexer_proptest.rs to only contain actual
property-based tests, moving non-proptest tests to more appropriate locations.

## Changes

### Created tests/lexer_samples.rs
- Moved 8 sample document snapshot tests from lexer_proptest.rs
- These are integration tests using insta snapshots, not property tests
- Tests verify lexer behavior against real sample files

### Added tests to src/lex/lexing.rs
- Moved 5 integration tests (test_paragraph_pattern, test_list_pattern, etc.)
- These test exact token sequences for specific patterns
- Belong in the lexing module since they test the lex() function directly

### Updated tests/lexer_proptest.rs
- Now contains ONLY property-based tests (8 proptest tests)
- Removed sample_document_tests module (moved to lexer_samples.rs)
- Removed integration_tests module (moved to lexing.rs)
- Added cross-references in doc comments to help locate related tests

### Snapshot Migration
- Moved 8 snapshot files from lexer_proptest__sample_document_tests__*
  to lexer_samples__*
- Updated snapshot paths to match new test file structure

## Rationale

The original file mixed three distinct test types:
1. Property-based tests (should stay) ✅
2. Sample document snapshot tests (integration tests, belong in tests/)
3. Exact pattern integration tests (unit tests, belong near code being tested)

This reorganization improves:
- **Clarity**: File names match content (proptest only has proptests)
- **Discoverability**: Tests are near the code they test
- **Organization**: Different test types in appropriate locations

## Test Results

All 589 tests passing:
- 8 property-based tests in lexer_proptest.rs
- 8 sample document tests in lexer_samples.rs
- 5 pattern integration tests in lexing.rs
- All other existing tests continue to pass
